edit unit name modal looks good now
