+++ client/src/scss/index.scss
@@ -18,6 +18,7 @@
 @import '../components/PerformaceFeedback/TeacherPerformance/teacherPerformance';
 @import '../components/PopUpForm/popupform';
 @import '../components/PageHeader/pageheader';
+@import '../components/EditCompanyNameModal/editcompanynamemodal';
 @import '../pages/./TestPage/testpage';
 @import '../pages/./LoginPage/loginpage';
 @import '../pages/LandingPage/landingpage';
